+++ changelog.md
@@ -1,5 +1,13 @@
 # Change Log
 ---
+[0.5.0]
+---
+## Breaking Change
+- test() concept has been removed, library will focus
+  on validation() concept which throw error object.
+  Focus in modifying error object to be more meaningful errors.
+  Removed : `Model.test()`, `Array.test()`
+- module `field.FieldError` has been chaged to `field.ValidateError`
 
 [0.4.0]
 ---